auth complited

+++ src/tasks/tasks.service.ts
@@ -22,8 +22,11 @@ export class TasksService {
 	// 	return this.tasks;
 	// }
 
-	async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-		return this.taskRepository.getTasks(filterDto);
+	async getTasks(
+		filterDto: GetTaskFilterDto,
+		user: User
+		): Promise<Task[]> {
+		return this.taskRepository.getTasks(filterDto, user);
 	}
 
 	// getTaskWithFilter(filterDto: GetTaskFilterDto): Task[] {
@@ -53,8 +56,11 @@ export class TasksService {
 	// 	return found;
 	// }
 
-	async getTaskById(id: number): Promise<Task> {
-		const found = await this.taskRepository.findOne(id);
+	async getTaskById(
+		id: number,
+		user: User
+		): Promise<Task> {
+		const found = await this.taskRepository.findOne({ where: { id, userId: user.id }});
 
 		if (!found) {
 			throw new NotFoundException(`Task with id = ${id} not found`);
@@ -90,8 +96,12 @@ export class TasksService {
 	// 	return found;
 	// }
 
-	async deleteTask(id: number): Promise<string> {
-		const result = await this.taskRepository.delete(id);
+	async deleteTask(
+		id: number,
+		user: User
+		): Promise<string> {
+		const userId = user.id.toString();
+		const result = await this.taskRepository.delete({ id, userId });
 
 		if (result.affected === 0) {
 			throw new NotFoundException(`Task with id = ${id} doesn't exist`);
@@ -100,8 +110,8 @@ export class TasksService {
 		}
 	}
 
-	async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-		const task = await this.getTaskById(id);
+	async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+		const task = await this.getTaskById(id, user);
 		task.status = status;
 		await task.save();
